% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{c_to_f}
\alias{c_to_f}
\title{Celcius to Far}
\usage{
c_to_f(temp_c)
}
\arguments{
\item{temp_c}{a temperature to convert}
}
\value{
a temp in f
}
\description{
Other way round from previous one
}
\examples{
c_to_f(1)

}
