% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{f_to_c}
\alias{f_to_c}
\title{Fahrenheit to Celsius}
\usage{
f_to_c(temp_f)
}
\arguments{
\item{temp_f}{a temperature to convert}
}
\value{
numeric of temperature in Celsius
}
\description{
Convert temperatures in degrees Fahrenheit to Celsius
}
\examples{
f_to_c(40)

}
